refactor accumulator
